+++ example/babel.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  presets: ['@zhihu/babel-preset/app'],
-  plugins: ['react-hot-loader/babel'],
-}