+++ example/main/index.js
@@ -1,7 +1,6 @@
 /* eslint-disable import/named */
 import '@babel/polyfill'
 import 'raf/polyfill'
-import 'whatwg-fetch'
 import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './App'