+++ packages/griffith-hls/src/index.js
@@ -1,1 +1,7 @@
-export {default} from './Video'
+import VideoComponent from './Video'
+
+export default {
+  pluginName: 'griffith-hls',
+  VideoComponent,
+  willHandleSrcChange: true,
+}