+++ packages/griffith-mp4/src/fetch.js
@@ -3,33 +3,41 @@ export default class FragmentFetch {
 
   constructor(url, start, end, callback) {
     this.start = start
-    this.controller = new AbortController()
     FragmentFetch.queue.push(this)
-    return fetch(url, {
-      headers: {
-        Range: `bytes=${start}-${end}`,
-      },
-      signal: this.controller.signal,
-    })
-      .then(res => {
-        callback(res.arrayBuffer())
-        this.remove()
-      })
-      .catch(() => {
-        this.remove()
-      })
+
+    const xhr = new XMLHttpRequest()
+    this.xhr = xhr
+    xhr.open('get', url)
+    xhr.responseType = 'arraybuffer'
+    xhr.setRequestHeader('Range', `bytes=${start}-${end}`)
+    xhr.setRequestHeader('Access-Control-Allow-Methods', 'GET, OPTIONS')
+    xhr.setRequestHeader('Access-Control-Allow-Origin', '*')
+    xhr.onload = () => {
+      if (xhr.status === 200 || xhr.status === 206) {
+        callback(xhr.response)
+      }
+      this.remove()
+    }
+    xhr.send()
+
+    xhr.onerror = () => {
+      this.remove()
+    }
+    xhr.onabort = () => {
+      this.remove()
+    }
   }
 
   remove = () => {
     FragmentFetch.queue = FragmentFetch.queue.filter(
-      item => item.start !== this.statr
+      item => item.start !== this.start
     )
   }
 
   static clear() {
     while (FragmentFetch.queue.length) {
       const item = FragmentFetch.queue.shift()
-      item.controller.abort()
+      item.xhr.abort()
     }
   }
 }