+++ packages/griffith-mp4/src/index.js
@@ -1,3 +1,7 @@
-/* eslint-disable import/default */
-import Player from './player'
-export default Player
+import VideoComponent from './player'
+
+export default {
+  pluginName: 'griffith-mp4',
+  VideoComponent,
+  willHandleSrcChange: true,
+}