+++ example/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.2](https://github.com/zhihu/griffith/compare/v1.4.1...v1.4.2) (2019-04-17)
+
+**Note:** Version bump only for package example
+
+
+
+
+
+## [1.4.1](https://github.com/zhihu/griffith/compare/v1.4.0...v1.4.1) (2019-04-16)
+
+**Note:** Version bump only for package example
+
+
+
+
+
+# [1.4.0](https://github.com/zhihu/griffith/compare/v1.3.1...v1.4.0) (2019-04-11)
+
+**Note:** Version bump only for package example
+
+
+
+
+
+## [1.3.1](https://github.com/zhihu/griffith/compare/v1.3.0...v1.3.1) (2019-04-03)
+
+
+### Bug Fixes
+
+* use xhr instead of the fetch api ([#56](https://github.com/zhihu/griffith/issues/56)) ([267b56a](https://github.com/zhihu/griffith/commit/267b56a))
+
+
+
+
+
+# [1.3.0](https://github.com/xiaoyuhen/griffith/compare/v1.2.1...v1.3.0) (2019-04-02)
+
+
+### Bug Fixes
+
+* last GOP dont have audio track ([#51](https://github.com/xiaoyuhen/griffith/issues/51)) ([67d37bb](https://github.com/xiaoyuhen/griffith/commit/67d37bb))
+
+
+
+
+
+## [1.2.1](https://github.com/xiaoyuhen/griffith/compare/v1.2.0...v1.2.1) (2019-03-27)
+
+**Note:** Version bump only for package example
+
+
+
+
+
 # [1.2.0](https://github.com/zhihu/griffith/compare/v1.1.4...v1.2.0) (2019-03-26)
 
 **Note:** Version bump only for package example