+++ packages/griffith-hls/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.4.0](https://github.com/zhihu/griffith/compare/v1.3.1...v1.4.0) (2019-04-11)
+
+
+### Bug Fixes
+
+* Fix a hls bug when switching quality ([#59](https://github.com/zhihu/griffith/issues/59)) ([4a0e079](https://github.com/zhihu/griffith/commit/4a0e079))
+
+
+### Features
+
+* adjust video quality ([#65](https://github.com/zhihu/griffith/issues/65)) ([f34077c](https://github.com/zhihu/griffith/commit/f34077c))
+
+
+
+
+
+## [1.2.1](https://github.com/xiaoyuhen/griffith/compare/v1.2.0...v1.2.1) (2019-03-27)
+
+**Note:** Version bump only for package griffith-hls
+
+
+
+
+
 # [1.2.0](https://github.com/zhihu/griffith/compare/v1.1.4...v1.2.0) (2019-03-26)
 
 **Note:** Version bump only for package griffith-hls