+++ packages/griffith-message/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/xiaoyuhen/griffith/compare/v1.2.0...v1.2.1) (2019-03-27)
+
+**Note:** Version bump only for package griffith-message
+
+
+
+
+
 # [1.2.0](https://github.com/zhihu/griffith/compare/v1.1.4...v1.2.0) (2019-03-26)