+++ packages/griffith-mp4/CHANGELOG.md
@@ -3,6 +3,90 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.2](https://github.com/zhihu/griffith/compare/v1.4.1...v1.4.2) (2019-04-17)
+
+
+### Bug Fixes
+
+* cannot read property handlerType of undefined ([#70](https://github.com/zhihu/griffith/issues/70)) ([3702079](https://github.com/zhihu/griffith/commit/3702079))
+* failed to execute endOfStream on MediaSource ([#69](https://github.com/zhihu/griffith/issues/69)) ([c8c91ff](https://github.com/zhihu/griffith/commit/c8c91ff))
+* failed to execute start on TimeRanges ([#71](https://github.com/zhihu/griffith/issues/71)) ([b59c4d0](https://github.com/zhihu/griffith/commit/b59c4d0))
+
+
+
+
+
+## [1.4.1](https://github.com/zhihu/griffith/compare/v1.4.0...v1.4.1) (2019-04-16)
+
+
+### Bug Fixes
+
+* add Access-Control-Allow-Methods limit ([4049347](https://github.com/zhihu/griffith/commit/4049347))
+* cannot read property hdlr of undefined ([#66](https://github.com/zhihu/griffith/issues/66)) ([038ec8b](https://github.com/zhihu/griffith/commit/038ec8b))
+* add Access-Control-Allow-Origin option ([#68](https://github.com/zhihu/griffith/issues/68)) ([8efb682](https://github.com/zhihu/griffith/pull/68/commits/8efb682)))
+* failed to execute appendBuffer on SourceBuffer ([#68](https://github.com/zhihu/griffith/issues/68)) ([51e6602](https://github.com/zhihu/griffith/pull/68/commits/51e6602)))
+
+
+
+
+# [1.4.0](https://github.com/zhihu/griffith/compare/v1.3.1...v1.4.0) (2019-04-11)
+
+
+### Bug Fixes
+
+* cache data ([#61](https://github.com/zhihu/griffith/issues/61)) ([0b5e874](https://github.com/zhihu/griffith/commit/0b5e874))
+* Fix a hls bug when switching quality ([#59](https://github.com/zhihu/griffith/issues/59)) ([4a0e079](https://github.com/zhihu/griffith/commit/4a0e079))
+* remove mseUpdating flag ([#63](https://github.com/zhihu/griffith/issues/63)) ([93867af](https://github.com/zhihu/griffith/commit/93867af))
+* replay case ([#60](https://github.com/zhihu/griffith/issues/60)) ([e2693ad](https://github.com/zhihu/griffith/commit/e2693ad))
+
+
+### Features
+
+* adjust video quality ([#65](https://github.com/zhihu/griffith/issues/65)) ([f34077c](https://github.com/zhihu/griffith/commit/f34077c))
+
+
+
+
+
+## [1.3.1](https://github.com/zhihu/griffith/compare/v1.3.0...v1.3.1) (2019-04-03)
+
+
+### Bug Fixes
+
+*  Failed to execute appendBuffer on SourceBuffer ([#57](https://github.com/zhihu/griffith/issues/57)) ([f4c435f](https://github.com/zhihu/griffith/commit/f4c435f))
+* use xhr instead of the fetch api ([#56](https://github.com/zhihu/griffith/issues/56)) ([267b56a](https://github.com/zhihu/griffith/commit/267b56a))
+
+
+
+
+
+# [1.3.0](https://github.com/xiaoyuhen/griffith/compare/v1.2.1...v1.3.0) (2019-04-02)
+
+
+### Bug Fixes
+
+*  Cannot read property appendBuffer of undefined ([#48](https://github.com/xiaoyuhen/griffith/issues/48)) ([9415b63](https://github.com/xiaoyuhen/griffith/commit/9415b63))
+* empty audio track ([#52](https://github.com/xiaoyuhen/griffith/issues/52)) ([a62ef4d](https://github.com/xiaoyuhen/griffith/commit/a62ef4d))
+* last GOP dont have audio track ([#51](https://github.com/xiaoyuhen/griffith/issues/51)) ([67d37bb](https://github.com/xiaoyuhen/griffith/commit/67d37bb))
+
+
+### Features
+
+* add buffered cache ([#43](https://github.com/xiaoyuhen/griffith/issues/43)) ([c7a7b2a](https://github.com/xiaoyuhen/griffith/commit/c7a7b2a))
+* multiple sourcebuffer ([#49](https://github.com/xiaoyuhen/griffith/issues/49)) ([ac64d8a](https://github.com/xiaoyuhen/griffith/commit/ac64d8a))
+
+
+
+
+
+## [1.2.1](https://github.com/xiaoyuhen/griffith/compare/v1.2.0...v1.2.1) (2019-03-27)
+
+**Note:** Version bump only for package griffith-mp4
+
+
+
+
+
 # [1.2.0](https://github.com/zhihu/griffith/compare/v1.1.4...v1.2.0) (2019-03-26)
 
 **Note:** Version bump only for package griffith-mp4